fix: correct asymmetric tail deflection flag
Bug in wire output from Vorlax led to incorrect wireframe plots, which applied AINC the same (positive AINC yields TE up) on surfaces defined in both the positive y and the negative y direction. The aerodynamic calcs do NOT reflect this, but instead reflect the following (and echoed on page A-18 of the NASA CR):
Positive AINC yields TE up (wash-out) on a surface defined in the positive y direction;
Positive AINC yields TE down (wash-in) on a surface defined in the negative y direction.
The mirror (left) tail in the discoidal vorModel is always defined in the negative y direction (the only surface defined in negative y -- the rest are defined in positive y and mirrored). Thus, the left tail requires special treatment of the anti-symmetric flag (namely, multiplication by -1) to ensure the intended deflection. Until the wireframe output is fixed in Vorlax (added as issue), only the aerodynamic results will reflect this reality (check runs of vorPlotLong with symmetric and asymmetric deflections to confirm).